requirements.txt
Django==3.1
djangorestframework==3.11.1
django-cleanup==5.0.0